long instead of int

+++ src/asm.h
@@ -5,6 +5,6 @@ int asm_add(int x, int y);
 long asm_add_seq(int n);
 long asm_fib(int n);
 long asm_fnv_hash(const char* str, int len);
-void asm_sort(int* lis, int len);
+void asm_sort(long* lis, long len);
 
 #endif
