+++ tests/test_asm.c
@@ -48,14 +48,14 @@ TEST("Test asm fnc hash") {
 
 
 TEST("Test asm sort function") {
-  int lis[] = {2, 1, 4, 3};
+  long lis[] = {2, 1, 4, 3};
   /*int expected[] = {1, 2, 3, 4};*/
   asm_sort(lis, 4);
   
   int i = 0;
   for(; i < 4; i++){
     /*ASSERT(lis[i], expected[i]);*/
-    printf("%i ", lis[i]);
+    printf("%i ", (int)lis[i]);
   }
